+++ frontend/public/about.html
@@ -11,7 +11,7 @@
       }
     </style>
     <link rel="stylesheet" href="./index.css" />
-    <script src="./about.js" defer></script>
+    <script src="./about.js" type="module" defer></script>
     <script src="https://api.vvhan.com/api/script/yinghua" defer></script>
   </head>
   <body>
